Add files via upload
Basic app based on React: Using State with Form Inputs, working on a text messaging app.
Exercise: Using State with Form Inputs
You're working on a text messaging app and your task is to validate the text entered by a user whilst the user is typing.
If the text message entered is valid (for this example: 
if it's at least 3 characters long), the text "Valid message" should be displayed below the input field. 
If it's invalid (i.e., shorter than 3 characters), the text "Invalid message" should be displayed.